Move HTTP request methods into new `HttpClient` class
